requirements.txt
fastapi==0.115.14
mangum==0.19.0
uvicorn==0.34.3
pydantic==2.9.2
aws-cdk-lib==2.202.0
httpx==0.28.1
python-dotenv==1.1.0
langchain-core==0.3.65
langchain==0.3.25
langchain-anthropic==0.3.3
langchain-community==0.3.24
langgraph==0.2.45
anthropic==0.55.0
